+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # 1.0.0 
 - Ability to use the same class multiple time
-- use with svelte loader Ok
+- Use with svelte loader Ok
 
 # 0.1.1
 - Fix modules exports